+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Improved UCC additional_packaging.py file for reduced manual code writing for UCC based Add-on code.
 * Improved debugging logs in app build generation.
+* Added `is_remove_mypyc_from_ucc_lib` input (default: `true`) to control removal of `*mypyc.cpython-*.so` files from UCC-generated `lib/` before packaging. This will help with Splunk App-inspect failures.
 
 ### Fixed